Added part to check for interfaces.
not sure why but the interface is always eth3 even if it was sent to another one of our interfaces. So I added this piece of code and a function in sr_if.c that searchs for an interface through the ip address and switches it with the current iface if it exists. not sure if this is what we should do, but now ping and traceroute works for the other interfaces too.